# perf: implement Ruffini rule division for 'x - b' monomials