Use a special-cased division for `P / (x - b)`. It's 20x faster than running the regular
division for this hotspot in provers.
Added property testings to ensure it's equivalent to the regular division.

## Type of change

- [x] New feature
- [x] Optimization

## Checklist
- [x] Unit tests added
- [x] This change requires new documentation.
  - [x] Documentation has been added/updated.
- [x] This change is an Optimization
  - [x] Benchmarks added/run

## Results on M1

```
Polynomial/div by 'x - b' with generic div
                        time:   [810.69 ns 821.65 ns 835.76 ns]
Polynomial/div by 'x - b' with Ruffini
                        time:   [21.992 ns 22.257 ns 22.496 ns]
```